Continuing the code fixes
